refactor: rename dark mode component
