+++ src/assets/styles/scaffold.css
@@ -1,7 +1,3 @@
-*,  
-*:before, 
-*:after { box-sizing: inherit; }
-
 html {
   font-size: var(--font-size-root);
   min-width: 300px;
@@ -9,7 +5,6 @@ html {
   text-rendering: optimizeLegibility;
   text-size-adjust: 100%;
   overflow-x: hidden;
-  box-sizing: border-box;
   -moz-osx-font-smoothing: grayscale;
   -webkit-font-smoothing: antialiased;
 }
@@ -44,7 +39,9 @@ body.dark-mode img {
   transition: opacity var(--transition);
 }
 
-body.dark-mode img:hover { opacity: 1; }
+@media (hover: hover) {
+  body.dark-mode img:hover { opacity: 1; }
+}
 
 body.dark-mode .social-logo {
   filter: invert(1);
