+++ src/components/shared/set-theme/SetThemeEnum.ts
@@ -1,4 +1,4 @@
-export enum DarkModeEnum {
+export enum SetThemeEnum {
   dark = 'dark',
   light = 'light'
 }
